Refactored the video code into RemoteControlManager
